+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 2.6.0 (unreleased)
 ------------------
 
+- #2540 Migrate Container Types to Dexterity
+- #2542 Migrate Manufacturers to Dexterity
+- #2547 Add get_user_email function in the api
+- #2544 Display creator fullname in samples listing
+- #2546 Add move_object function in the API
+- #2539 Add User Profile / Password Reset
+- #2543 Fix AttributeError for Instrument Adapters
 - #2533 Migrate Sample Points to Dexterity
 - #2537 Support multi-line text on result entry
 - #2536 Fix counts from control-panel includes client-specific items